producer/requirements.txt
django
djangorestframework
celery
redis
amqp
mysql